store configuration as vec4 instead of uvec4
this avoids having to use the conversion functions

+++ filament/src/fsr.h
@@ -39,10 +39,10 @@ struct FSRSharpeningConfig {
 };
 
 struct FSRUniforms {
-    math::uint4 EasuCon0;
-    math::uint4 EasuCon1;
-    math::uint4 EasuCon2;
-    math::uint4 EasuCon3;
+    math::float4 EasuCon0;
+    math::float4 EasuCon1;
+    math::float4 EasuCon2;
+    math::float4 EasuCon3;
     math::uint4 RcasCon;
 };
